requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
langchain-community>=0.0.20
langchain-core>=0.3.49
langgraph>=0.0.25
qdrant-client==1.6.4
openai>=1.68.2
langsmith>=0.0.77
pydantic>=2.0.0
langchain-openai>=0.3.12

